Refactor: 파일 업로드 서버 연결
